[Refactor] common AutoCompleteCompanyNameTextInput (#788)
* common/autocomplete-text-input_new

* company-name-auto-complete-text-input
