+++ src/components/TimeAndSalary/SearchBar.js
@@ -1,16 +1,14 @@
-import React, { useState, useCallback, useRef } from 'react';
+import React, { useState, useCallback } from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
 import R from 'ramda';
 import ReactPixel from 'react-facebook-pixel';
 import { withRouter } from 'react-router-dom';
 
-import { debounce } from 'utils/streamUtils';
-import AutoCompleteTextInput from './AutoCompleteTextInput';
+import AutoCompleteCompanyNameTextInput from 'common/form/AutoCompleteTextInput_new/AutoCompleteCompanyNameTextInput';
 import Magnifiner from 'common/icons/Magnifiner';
 
 import styles from './SearchBar.module.css';
-import { fetchCompanyCandidates } from '../../apis/timeAndSalaryApi';
 import { searchKeywordSelector } from './common/selectors';
 
 import PIXEL_CONTENT_CATEGORY from '../../constants/pixelConstants';
@@ -25,39 +23,6 @@ const searchType = 'company';
 const SearchBar = ({ history, location }) => {
   const [searchText, setSearchText] = useState(
     getInitialSearchTextFromLocation({ location }),
-  );
-  const [autocompleteItems, setAutocompleteItems] = useState([]);
-  const eleRef = useRef(null);
-
-  const performSearch = useCallback(
-    debounce(async searchText => {
-      if (searchText) {
-        try {
-          const response = await fetchCompanyCandidates({ key: searchText });
-          const autocompleteItems = response.map(({ _id: { name } }) => name);
-          if (eleRef.current) {
-            setAutocompleteItems(autocompleteItems);
-          }
-        } catch (err) {
-          if (eleRef.current) {
-            setAutocompleteItems([]);
-          }
-        }
-      } else {
-        if (eleRef.current) {
-          setAutocompleteItems([]);
-        }
-      }
-    }, 500),
-    [setAutocompleteItems],
-  );
-
-  const handleSearchTextChange = useCallback(
-    e => {
-      setSearchText(e.target.value);
-      performSearch(e.target.value);
-    },
-    [performSearch],
   );
 
   const gotoSearchResult = useCallback(
@@ -75,7 +40,7 @@ const SearchBar = ({ history, location }) => {
     [history],
   );
 
-  const handleAutocompleteItemSelect = useCallback(
+  const handleAutocompleteItemSelected = useCallback(
     e => {
       gotoSearchResult(e);
     },
@@ -92,18 +57,16 @@ const SearchBar = ({ history, location }) => {
 
   return (
     <form
-      ref={eleRef}
       className={cn(styles.section, styles.searchbar)}
       onSubmit={handleFormSubmit}
     >
-      <AutoCompleteTextInput
+      <AutoCompleteCompanyNameTextInput
         wrapperClassName={styles.textInputWrapper}
         className={styles.textInput}
         value={searchText}
-        onChange={handleSearchTextChange}
+        onChange={setSearchText}
         placeholder="輸入公司、職稱查詢"
-        autocompleteItems={autocompleteItems}
-        onAutocompleteItemSelect={handleAutocompleteItemSelect}
+        onCompanyNameSelected={handleAutocompleteItemSelected}
       />
       <button type="submit" className={styles.searchBtn}>
         <Magnifiner />
